requirements.txt
# API GraphQL para la BDNS - Requisitos

# API
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
strawberry-graphql>=0.171.1
pydantic>=1.10.7
pydantic-settings>=2.0.0

# Base de datos
sqlalchemy>=2.0.23
alembic>=1.10.3
oracledb>=1.3.1
psycopg2-binary>=2.9.6

# ETL
pandas>=2.0.0
requests>=2.31.0
tenacity>=8.2.2
apscheduler>=3.10.1

# Caché
redis>=4.5.4
aiocache>=0.12.1

# Monitorización
prometheus-client>=0.16.0
structlog>=23.1.0

# Utilidades
python-dotenv>=1.0.0
openpyxl>=3.1.2
pandas >=2.0.0